+++ example/a.ts
@@ -1,15 +1,21 @@
 import { Elysia, t } from '../src'
 import { req } from '../test/utils'
 
+const delay = (time = 1000) => new Promise((r) => setTimeout(r, time))
+
 const app = new Elysia()
-	.mapResponse(async ({ response }) => {})
-	.get('/', async () => 'NOOP')
-	.compile()
+	.trace(async ({ beforeHandle, handle, afterHandle, set }) => {
+		await beforeHandle
+		await handle
+		await afterHandle
 
-const res = await app.handle(req('/')).then((x) => x.text())
-
-app.handle(req('/'))
-	.then((x) => x.text())
-	.then(console.log)
-
-// console.log(app.routes[0]?.composed?.toString())
+		set
+	})
+	.get('/', () => 'A', {
+		beforeHandle() {
+			return 'a'
+		},
+		afterHandle() {}
+	})
+	.get('static', Bun.file('test/kyuukurarin.mp4'))
+	.listen(3000)